+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `soulcodex/keyable` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2021-02-24
 
 - initial release
+- forked from givebutter/laravel-keyable
+- added support for multiple types of api key access
+- added uuid support for api keys table